Store healing spells by type
Healing spells are sorted and stored into 14 categories.

Added HealAmount() and PowerCost() methods to help with healing spell sorting and selection.

Added OnCastSpellLosCheckFail() to move towards a spell target if it's out of LOS.  This stops a rare instance in which a mimic would stand in one place and spam spells at a target it doesn't have LOS to.