Added method to check group health/status
Added a method to check the group health/status, and save the results in a series of properties so it only checks once per think interval.

Modified SetMainPuller() and SetPullPoint() to allow them to reset their respective values.

Modified to send messages to the group so other players can see how bots are set up.