Added and Modified Commands
Added /mheal to set healer status

Added /mpull to set camp, set pull, and pull in a single command

Added /mfollow to clear camp/pull and set mimics to follow state

Added /mattack to have mimics attack a target, and shift targets if already in combat

Added /mintercept and /mprotect to have mimics use intercept and protect abilities

Changed /m to /mcreate to make it more clear.  Added arguments to invite the new mimic to group and to specify spec.

Added a "here" option for /mcamp

Fixed an issue with /msummon breaking click on group window functionality when changing zones